+++ src/main/java/com/cpsc304/HotelManagement/Model/RoomRecord.java
@@ -3,12 +3,12 @@ package com.cpsc304.HotelManagement.Model;
 import java.util.Date;
 
 public class RoomRecord {
-    Long rm_number;
+    Integer rm_number;
     Integer price;
     Date date;
-    Long last_req;
+    Integer last_req;
 
-    public RoomRecord(Long rm_number, Integer price, Date date, Long last_req) {
+    public RoomRecord(Integer rm_number, Integer price, Date date, Integer last_req) {
         this.rm_number = rm_number;
         this.price = price;
         this.date = date;
@@ -18,11 +18,11 @@ public class RoomRecord {
     public RoomRecord() {
     }
 
-    public Long getRm_number() {
+    public Integer getRm_number() {
         return rm_number;
     }
 
-    public void setRm_number(Long rm_number) {
+    public void setRm_number(Integer rm_number) {
         this.rm_number = rm_number;
     }
 
@@ -42,11 +42,11 @@ public class RoomRecord {
         this.date = date;
     }
 
-    public Long getLast_req() {
+    public Integer getLast_req() {
         return last_req;
     }
 
-    public void setLast_req(Long last_req) {
+    public void setLast_req(Integer last_req) {
         this.last_req = last_req;
     }
 }
